+++ mobile/apps/photos/lib/ui/settings/support/help_support_page.dart
@@ -152,7 +152,7 @@ class HelpSupportPage extends StatelessWidget {
     final colorScheme = getEnteColorScheme(context);
     return HugeIcon(
       icon: icon,
-      color: colorScheme.strokeBase,
+      color: colorScheme.menuItemIconStroke,
       size: 20,
     );
   }